requirements.txt
torch
torchvision
streamlit
numpy
opencv-python-headless
psutil
scipy
pillow
